+++ packages/animated-pose/src/inc/default-transitions.ts
@@ -1,49 +1,27 @@
-import { Transition } from '../types';
-import { Animated } from 'react-native';
-import { eachValue } from './transition-composers';
+const tween = {
+  duration: 300
+};
 
-const tween: Transition = ({ value, toValue, useNativeDriver = true }) =>
-  Animated.timing(value, {
-    toValue,
-    duration: 300,
-    useNativeDriver
-  });
+const linearTween = {
+  ...tween,
+  easing: (v: number) => v
+};
 
-const linearTween: Transition = ({ value, toValue, useNativeDriver = true }) =>
-  Animated.timing(value, {
-    toValue,
-    duration: 300,
-    easing: v => v,
-    useNativeDriver
-  });
+const overDampedSpring = ({ toValue }: { toValue: number }) => ({
+  type: 'spring',
+  stiffness: 700,
+  damping: toValue === 0 ? 100 : 35
+});
 
-const overDampedSpring: Transition = ({
-  value,
-  toValue,
-  useNativeDriver = true
-}) =>
-  Animated.spring(value, {
-    toValue,
-    stiffness: 700,
-    damping: toValue === 0 ? 100 : 35,
-    useNativeDriver
-  });
+const underDampedSpring = {
+  type: 'spring',
+  stiffness: 500,
+  damping: 25,
+  restDelta: 0.5,
+  restSpeed: 10
+};
 
-const underDampedSpring: Transition = ({
-  value,
-  toValue,
-  useNativeDriver = true
-}) =>
-  Animated.spring(value, {
-    toValue,
-    stiffness: 500,
-    damping: 25,
-    restDisplacementThreshold: 0.5,
-    restSpeedThreshold: 10,
-    useNativeDriver
-  });
-
-const intelligentTransition = eachValue({
+const intelligentTransition = {
   x: underDampedSpring,
   y: underDampedSpring,
   z: underDampedSpring,
@@ -57,6 +35,6 @@ const intelligentTransition = eachValue({
   scale: overDampedSpring,
   opacity: linearTween,
   default: tween
-});
+};
 
 export default new Map([['default', intelligentTransition]]);