+++ packages/react-pose-core/src/components/Transition/types.ts
@@ -1,0 +1,19 @@
+import * as React from 'react';
+
+export type Props = {
+  children?: React.ReactElement<any> | Array<React.ReactElement<any>>;
+  preEnterPose?: string;
+  enterPose?: string;
+  exitPose?: string;
+  animateOnMount?: boolean;
+  enterAfterExit?: boolean;
+  popFromLayoutOnExit?: boolean;
+  [key: string]: any;
+};
+
+export type State = {
+  children: Array<React.ReactElement<any>>;
+  leaving: { [key: string]: boolean };
+  hasMounted: boolean;
+  scheduleChildRemoval: (key: string) => void;
+};