+++ packages/react-pose-core/src/index.ts
@@ -1,2 +1,5 @@
 import posed from './posed';
+import Transition from './components/Transition';
+
 export default posed;
+export { Transition };