+++ packages/animated-pose/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 Animated Pose adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] 2018-09-14
+
+### Added
+
+- `keyframes` animation.
+
+## [1.1.0] 2018-09-13
+
+### Added
+
+- Support for transition definitions.
+
+## [1.0.0] 2018-09-12
+
+### Upgrade
+
+- `pose-core@2.0.0`.
+
 ## [0.3.4] 2018-06-15
 
 ### Adding