+++ packages/react-native-pose/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 React Native Pose adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.0] 2018-09-14
+
+### Upgraded
+
+- `react-pose-core@0.4.0`
+
+## [0.7.0] 2018-09-14
+
+### Upgraded
+
+- `animated-pose@1.2.0`
+
+## [0.6.0] 2018-09-13
+
+### Upgraded
+
+- `animated-pose@1.1.0`
+
+## [0.5.0] 2018-09-12
+
+### Upgraded
+
+- `animated-pose@1.0.1`
+
 ## [0.4.1] Unreleased
 
 ### Added